// packages/core/src/types.ts
export interface IDesignerProps {
  package?: string //npm包名
  registry?: string //web npm注册平台地址
  version?: string //semver版本号
  path?: string //组件模块路径
  title?: string //标题
  description?: string //描述
  icon?: string //icon
  droppable?: boolean //是否可作为拖拽容器，默认为true
  draggable?: boolean //是否可拖拽，默认为true
  deletable?: boolean //是否可删除，默认为true
  cloneable?: boolean //是否可拷贝，默认为true
  resizable?: IResizable
  translatable?: ITranslate // 自由布局
  inlineChildrenLayout?: boolean //子节点内联，用于指定复杂布局容器，强制内联
  selfRenderChildren?: boolean //是否自己渲染子节点
  propsSchema?: ISchema // JSON Schema
  defaultProps?: any //默认属性
  getDragNodes?: (node: TreeNode) => TreeNode | TreeNode[] //拦截转换Drag节点
  getDropNodes?: (node: TreeNode, parent: TreeNode) => TreeNode | TreeNode[] //拦截转换Drop节点
  getComponentProps?: (node: TreeNode) => any //拦截属性
  allowAppend?: (target: TreeNode, sources?: TreeNode[]) => boolean
  allowSiblings?: (target: TreeNode, sources?: TreeNode[]) => boolean
  allowDrop?: (target: TreeNode) => boolean
  [key: string]: any
}

export type IDesignerPropsMap = Record<string, IDesignerProps>

export type IDesignerControllerProps =
  | IDesignerProps
  | ((node: TreeNode) => IDesignerProps)

export type IDesignerControllerPropsMap = Record<
  string,
  IDesignerControllerProps
>
export interface IDesignerLocales {
  [ISOCode: string]: {
    [key: string]: any
  }
}

export interface IDesignerMiniLocales {
  [ISOCode: string]: string
}

export interface IDesignerBehaviors {
  [key: string]: IBehaviorHost
}

export interface IDesignerStore<P> {
  value: P
}

export type IDesignerIcons = Record<string, any>

export type IDesignerIconsStore = IDesignerStore<IDesignerIcons>
export type IDesignerLocaleStore = IDesignerStore<IDesignerLocales>
export type IDesignerBehaviorStore = IDesignerStore<IBehavior[]>
export type IDesignerLanguageStore = IDesignerStore<string>

// 工作台类型
export type WorkbenchTypes =
  | 'DESIGNABLE' // 编辑器
  | 'PREVIEW' // 预览器
  | 'JSONTREE' // json
  | 'MARKUP'
  | (string & {})

export interface IBehavior {
  name: string //组件名称
  extends?: string[] //行为扩展
  selector: (node: TreeNode) => boolean //选择器
  designerProps?: IDesignerControllerProps //行为属性
  designerLocales?: IDesignerLocales //行为国际化描述
}

export interface IBehaviorCreator {
  name: string //组件名称
  extends?: string[] //行为扩展
  selector: string | ((node: TreeNode) => boolean) //选择器
  designerProps?: IDesignerControllerProps //行为属性
  designerLocales?: IDesignerLocales //行为国际化描述
}

export interface IBehaviorHost {
  Behavior?: IBehavior[]
}

export type IBehaviorLike = IBehavior[] | IBehaviorHost